Adiçoes de polimorfismo e modo DEBUG implementado

+++ src/funcionarios/tratador.cpp
@@ -52,7 +52,7 @@ void Tratador::setUniforme(Uniforme u) {
 
 ostream& Tratador::printOutDados(ostream& o) const {
     int tamanhoColuna = 20, espacamento = 3;
-    int tamanhoTruncado = tamanhoColuna - 3 - espacamento;
+    long unsigned int tamanhoTruncado = tamanhoColuna - 3 - espacamento;
 
     o << setfill(' ') << setw(tamanhoColuna + espacamento) << ((this->getNome().length() > tamanhoTruncado) ? (this->getNome().substr(0, tamanhoTruncado) + "...") : this->getNome())
     << setfill(' ') << setw(tamanhoColuna + espacamento) << ((this->getEmail().length() > tamanhoTruncado) ? (this->getEmail().substr(0, tamanhoTruncado) + "...") : this->getEmail())
